change folder structure for start-up script
