+++ package.json
@@ -12,7 +12,7 @@
   "scripts": {
     "heroku-postbuild": "webpack -p --config ./webpack.config.js",
     "react-dev": "webpack -d --watch",
-    "start": "nodemon index.js",
+    "start": "nodemon server/index.js",
     "test": "mocha test/index.js"
   },
   "dependencies": {
