+++ server/index.js
@@ -1,44 +1,8 @@
-const express = require('express');
-const app = express();
-const bodyParser = require('body-parser');
-const sampledata = require('../sampledata.js');
+'use strict';
 
-// parse application/x-www-form-urlencoded
-app.use(bodyParser.urlencoded())
+var server = require('./app.js');
+var port = process.env.PORT || 3000;
 
-// parse application/json
-app.use(bodyParser.json())
-
-app.use(express.static(__dirname + '/../client/dist'));
-
-app.get('/feed/global', (req, res) => {
-
-  // DB query should replace default promise
-  Promise.resolve(sampledata.sampleFeed)
-    .then((results) => {
-      res.status(200).json({items: results});
-    })
-    .catch((err) => {
-      console.log(err);
-      res.sendStatus(404);
-    })
-
+server.listen(port, function () {
+  console.log('Server running on port %d', port);
 });
-
-app.get('/feed/user/:userId', (req, res) => {
-  
-  // DB query should replace default promise
-  let userId = req.params && req.params.userId;
-
-  Promise.resolve(sampledata.sampleFeed)
-    .then((results) => {
-      res.status(200).json({items: results});
-    })
-    .catch((err) => {
-      console.log(err);
-      res.sendStatus(404);
-    })
-});
-
-module.exports = app;
-
